Cleaned up code and added Dot Click through to dots fixing ISSUE [#14]
I used the ellipseRadiuses to detemine the displacment of both the dot and line to keep the propitions consistent even if the size of the original dot is changed in the future.
I used polar coordinates for the position of the additional line to make both adding the stright line on its end easier but also to remove chances of lines not going in a perfect circle.
Both components have been added as a group making them draggable together.